+++ tools/gen_documentation.sh
@@ -1,7 +1,9 @@
 #!/bin/bash -eu
 # Script for generating PDF versions of the documentation pages.
 
-WKHTMLTOPDF="${WKHTMLTOPDF:-$(which wkhtmltopdf)}"
+# Using which with a safe default so that this works in -u mode.
+WKHTMLTOPDF="${WKHTMLTOPDF:-$(which wkhtmltopdf || echo '')}"
+
 if [ ! -x "$WKHTMLTOPDF" ]; then
   >&2 echo 'Please install wkhtmltopdf (0.12.3 or newer) or set WKHTMLTOPDF to point to the binary.'
   exit 1
@@ -9,8 +11,8 @@ fi
 WKHTML_OPT='--lowquality --footer-center [page] --margin-top 20mm --margin-bottom 20mm'
 
 echo 'Starting LynxKite...'
-cd "$(dirname $0)/.."
-'./run.sh' &
+cd "$(dirname $0)/../web"
+gulp serve &
 LYNXKITE_PID=$!
 cd -
 function kill_grunt {
@@ -19,7 +21,7 @@ function kill_grunt {
 }
 trap kill_grunt EXIT
 # Wait until Grunt is up.
-while ! nc -z localhost 9090; do sleep 1; done
+tools/wait_for_port.sh 9090
 echo # Add new-line after Grunt output.
 
 echo 'Generating User Manual...'