+++ CHANGELOG.md
@@ -6,6 +6,24 @@ Please add changes to "master", preferably ordered by their significance. (Most 
 
 ### master
 
+### 1.8.0
+
+ - Major performance and scalability improvements.
+ - New option has been added in bucketed view: relative edge density.
+ - New operation "Find vertex coloring" has been added.
+ - Add search box to built-in help.
+ - Experimental feature: LynxKite can be used from [Jupyter](http://jupyter.org) (IPython Notebook).
+
+### 1.7.5
+
+ - Improve speed and stability of the project history editor
+ - Easier to find data export button
+ - Table import wizard: added tab stops and removed highlight flickering
+ - Hide ACL settings in single-user instances
+ - New aggregator: `median`.
+
+### 1.7.4
+
  - Prefix-based user access control can be specified in the prefix definitions file.
  - emr.sh: support setting up LynxKite in an Amazon VPC.
  - Numeric fields in CSV files can be imported to LynxKite with the right types.
@@ -17,6 +35,7 @@ Please add changes to "master", preferably ordered by their significance. (Most 
    vertices with no location from a map visualization.
  - Stability improvements regarding edge loading and handling graphs with large degree vertices.
  - SQL query results can be saved as segmentations.
+ - Prettier project history editor.
 
 ### 1.7.3